Development

+++ server/src/module/user/userController.js
@@ -1,25 +1,11 @@
 const userService = require('./userService')
 
-// GET all Userss
-const getAllUsers = async (req, res) => {
-  try {
-    const users = await userService.getAllUsers()
-    if(users.length == 0){
-      return res.status(404).json({message:'No data available at this time'})
-    }
-    return res.status(200).json(users)
-  } catch(error) {
-    return res.status(500).json({ error: error.message, stack: error.stack })
-  }
-}
-
 // GET User Page
 const getAllUserPage= async(req,res)=>{
   try{
-    const { page } = req.query;
-    const users = await userService.getAllUserPage(page)
+    const users = await userService.getAllUserPage(req.query)
     if(users.length == 0){
-      return res.status(404).json({message:'No data available at this time'})
+      return res.status(404).json({message:'No data available at this time'}) //change status to 200
     }
     return res.status(200).json(users)
   }catch(error)
@@ -31,8 +17,8 @@ const getAllUserPage= async(req,res)=>{
 // GET Users by ID
 const getUserById = async (req, res) => {
   try {
-    const user = await userService.getUser(req.params.id)
-    if (!user){
+    const user = await userService.getUserById(req.params.id)
+    if (user.length == 0){
       return res.status(404).json({ message: 'User not found' })
     } 
     return res.status(200).json(user)
@@ -44,10 +30,6 @@ const getUserById = async (req, res) => {
 // POST - Add new User
 const addUser = async (req, res) => {
   try {
-    const { name,surname, email, password,avatar,dateOfBirth } = req.body
-    if ( !surname || !email || !password ||!avatar || !dateOfBirth){
-      return res.status(400).json({ error: 'Name,Surname,email,avatar,dateofbirth and password are required' })
-    } 
     const newUser = await userService.createUser(req.body)
     return res.status(201).json(newUser)
   } catch(error) {
@@ -58,6 +40,9 @@ const addUser = async (req, res) => {
 // DELETE User by ID
 const deleteUser = async (req, res) => {
   try {
+    if(req.userId != req.params.id){
+      return res.status(403).json({message:'You can not delete others account'})
+    }
     const User = await userService.deleteUser(req.params.id)
     if (User.rowCount == 0){
       return res.status(404).json({ message: 'User not found' })
@@ -70,13 +55,11 @@ const deleteUser = async (req, res) => {
 
 // PATCH - Update User by ID
 const updateUser = async (req, res) => {
-  try {
-    const { id } = req.params
-    const { name, email } = req.body
-    if (!name || !email || !surname || !avatar){
-      return res.status(400).json({ error: 'Name,surname,avatar and email are required' })
-    }  
-    const updatedUser = await userService.updateUser(id, req.body)
+  try {  
+    if(req.userId != req.params.id){
+      return res.status(403).json({message:'You have not access this route'})
+    }
+    const updatedUser = await userService.updateUser(req.params.id, req.body)
     if(updatedUser.length == 0) {
       return res.status(404).json({ message: 'User not found' })
     }
@@ -86,4 +69,4 @@ const updateUser = async (req, res) => {
   }
 }
 
-module.exports = { getAllUsers, getUserById, addUser, deleteUser, updateUser,getAllUserPage }
+module.exports = {  getUserById, addUser, deleteUser, updateUser,getAllUserPage }
